requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
joblib==1.2.0
nltk==3.8.1
numpy==1.24.1
oauthlib==3.2.2
pandas==1.5.3
praw==7.6.1
prawcore==2.3.0
python-dateutil==2.8.2
pytz==2022.7.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
textblob==0.17.1
tqdm==4.64.1
tweepy==4.12.1
tzdata==2022.1
update-checker==0.18.0
urllib3==1.26.14
websocket-client==1.4.2
alpaca-py==0.7.1